authentication/requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.10
pyjwt==2.8.0
werkzeug==3.0.6
flask-cors==5.0.0